Fix a evil bug with a evil patch
I don't know why adding updateGrid(); fix one bug (on pre-gameOver)
but adds another: consecutives rows deletion after the first time
doesn't works.

For that I just create a variable private called `falls` which
counts the number of a block can falls on the board. If the first
fails is setted falls==0 so then the game cannot continue. The
gameOver screen is called.