Improve Game Over logic: remove falls counter and add insertOnGrid()
The previous logic was a problem because not always the first fall
failed is a GameOver. The player, on the end, can quickly just go
to left or right and wait for the next piece. This is allowed.

But using the counter falls for each fall and checking falls==0
for game Over is based. This is not allowed.

I'm using now the previous logic of gameOver but with one
differential: insertOnGrid is called on start() always when a new
group is instantiated and is valid. This make sure that even if the
block cannot fall at least once, it will inserted on the grid.

If this is not done, several pieces can over-leaped when the first fall
is always not allowed. Which makes a really strange buggy of the
player waiting for the game over and more pieces are created on the
fly.

In previous version I tried updated always the grid on start, but this
is bad because update grid delete previous childs. So I just splitted
updateGrid() part of grid insertion to insertOnGrid().